+++ src/main/java/rabun/oanda/rest/models/CandleMid.java
@@ -6,6 +6,6 @@ public class CandleMid {
     public float highMid;
     public float lowMid;
     public float closeMid;
-    public int volume;
+    public long volume;
     public boolean complete;
 }
