+++ src/main/java/rabun/oanda/rest/models/Instrument.java
@@ -4,5 +4,5 @@ public class Instrument {
     public String instrument;
     public String displayName;
     public float pip;
-    public int maxTradeUnits;
+    public long maxTradeUnits;
 }
