+++ src/main/java/rabun/oanda/rest/models/TradeOpened.java
@@ -1,6 +1,6 @@
 package rabun.oanda.rest.models;
 
 public class TradeOpened {
-    public int id;
-    public int units;
+    public long id;
+    public long units;
 }
