+++ src/main/java/rabun/oanda/rest/models/Transaction.java
@@ -1,8 +1,8 @@
 package rabun.oanda.rest.models;
 
 public class Transaction {
-    public int id;
-    public int accountId;
+    public long id;
+    public long accountId;
     public String time;
     public OandaTypes.TransactionType type;
