+++ src/main/java/rabun/oanda/rest/models/TransactionOrderFilled.java
@@ -1,9 +1,9 @@
 package rabun.oanda.rest.models;
 
 public class TransactionOrderFilled extends TransactionDefault {
-    public int pl;
-    public int interest;
-    public int orderId;
+    public long pl;
+    public long interest;
+    public long orderId;
     public TradeOpened tradeOpened;
 }
