+++ src/main/java/rabun/oanda/rest/models/TransactionOrderUpdate.java
@@ -1,9 +1,9 @@
 package rabun.oanda.rest.models;
 
 public class TransactionOrderUpdate extends Transaction {
-    public int units;
+    public long units;
     public float price;
     public OandaTypes.Reason reason;
-    public Integer expiry;
+    public long expiry;
 }
