+++ src/main/java/rabun/oanda/rest/models/TransactionTradeClose.java
@@ -3,7 +3,7 @@ package rabun.oanda.rest.models;
 
 public class TransactionTradeClose extends TransactionSimple {
     public float pl;
-    public int interest;
+    public long interest;
     public float accountBalance;
-    public int tradeId;
+    public long tradeId;
 }
