+++ src/main/java/rabun/oanda/rest/models/TransactionTradeUpdate.java
@@ -1,9 +1,9 @@
 package rabun.oanda.rest.models;
 
 public class TransactionTradeUpdate extends Transaction {
-    public int units;
+    public long units;
     public OandaTypes.Side side;
     public float stopLossPrice;
-    public int tradeId;
+    public long tradeId;
 
 }
